+++ package.json
@@ -4,8 +4,8 @@
   "description": "An Overwatch bot that manages 6v6-Events or Scrims.",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
-    "watch": "nodemon index.js",
+    "start": "node src/index.js",
+    "watch": "nodemon src/index.js",
     "test": "jest",
     "lint": "standard",
     "format": "standard --fix"
