+++ src/commands/clear.js
