+++ src/commands/configure.js
