+++ src/commands/endmatch.js
