+++ src/commands/endsession.js
