+++ src/commands/newsession.js
