+++ src/commands/randomize.js
