+++ src/commands/switchmode.js
